backend/requirements.txt
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
biopython==1.85
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.7.14
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
courlan==1.3.2
cymem==2.0.11
dateparser==1.2.2
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
elastic-transport==9.1.0
elasticsearch==9.1.0
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==3.2.0
hf-xet==1.1.5
hpack==3.0.0
hstspreload==2025.1.1
htmldate==1.9.3
httpcore==1.0.9
httpie==3.2.4
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.3
hyperframe==5.2.0
idna==2.10
it_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/it_core_news_lg-3.8.0/it_core_news_lg-3.8.0-py3-none-any.whl#sha256=b78582d0b2d05fd6509995f68ab7452efed7a27c6fbc5a071e9a9787a58c1e87
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jusText==3.0.2
langcodes==3.5.0
langid==1.1.6
language_data==1.3.0
lxml==5.4.0
lxml_html_clean==0.4.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
networkx==3.5
nltk==3.9.1
numpy==2.3.2
oauthlib==3.3.1
openai==1.99.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
praw==7.8.1
prawcore==2.4.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.8.0
pymetamap==0.1
pymongo==4.13.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytrends==4.9.2
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.31
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.1.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.34.2
sentence-transformers==5.1.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tld==0.13.1
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.54.1
trio==0.30.0
trio-websocket==0.12.2
tweepy==4.16.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
zhipuai==2.1.5.20250801
psutil==6.1.0